kernel: make AppId unique
AppId.idx used to refer to the index of the app in the process array.
However, if an app restarts then an "old" AppId that was created for the
process before it restarted will also work after an app is restarted,
but the restarted version may use entirely different resources. This
change changes the AppId after an app is restarted so that to the rest
of the OS it looks like an entirely different application.